feat: createdAt and updatedAt tags are assigned to todo items
When a new todo is created with the 'add' command, it's createdAt and updatedAt
tags will be assigned the current time. For every call to updateTodo, the
updatedAt tag will also be changed to the current time, to keep track of item
modifications. The id tag was also removed as it now serves no purpose.